+++ tests/commands/new-init-test.js
@@ -8,6 +8,8 @@ import temp from 'temp'
 
 let cli = require('../../src/cli')
 
+const DEP_VERSION_RE = /\d+\.\d+\.x/
+
 function reactAppAssertions(dir, name, err, done) {
   expect(err).toNotExist('No errors creating new React app')
   expect(glob.sync('**', {
@@ -35,6 +37,7 @@ function reactAppAssertions(dir, name, err, done) {
      .toInclude('node_modules/react-dom')
   let pkg = require(path.resolve(dir, 'package.json'))
   expect(pkg.name).toBe(name)
+  expect(pkg.devDependencies.nwb).toMatch(DEP_VERSION_RE)
   let config = require(path.resolve(dir, 'nwb.config.js'))
   expect(config).toEqual({type: 'react-app'})
   done()
@@ -68,6 +71,7 @@ function reactComponentAssertions(dir, name, err, done) {
   let pkg = require(path.resolve(dir, 'package.json'))
   expect(pkg.name).toBe(name)
   expect(pkg['jsnext:main']).toBe('es6/index.js')
+  expect(pkg.devDependencies.nwb).toMatch(DEP_VERSION_RE)
   let config = require(path.resolve(dir, 'nwb.config.js'))
   expect(config).toEqual({
     type: 'react-component',
@@ -100,6 +104,7 @@ function webAppAssertions(dir, name, err, done) {
   ])
   let pkg = require(path.resolve(dir, 'package.json'))
   expect(pkg.name).toBe(name)
+  expect(pkg.devDependencies.nwb).toMatch(DEP_VERSION_RE)
   let config = require(path.resolve(dir, 'nwb.config.js'))
   expect(config).toEqual({type: 'web-app'})
   done()
@@ -125,6 +130,7 @@ function webModuleAssertions(dir, name, err, done) {
   let pkg = require(path.resolve(dir, 'package.json'))
   expect(pkg.name).toBe(name)
   expect(pkg['jsnext:main']).toBe('es6/index.js')
+  expect(pkg.devDependencies.nwb).toMatch(DEP_VERSION_RE)
   let config = require(path.resolve(dir, 'nwb.config.js'))
   expect(config).toEqual({
     type: 'web-module',